requirements.txt
python-socketio
eventlet
requests
Flask
python-engineio
flask-cors
python-decouple
moment
redis
gql[all]